Add files via upload
ABSTRACT:


The outspread of the corona virus types has caused pandemic across the globe.The virus is
known for severe acute respiratory issues. The ailment it causes is called coronavirus disease
2019 or (COVID-19) affecting our day-to-day life disrupting world trade and movements.
Wearing a mask is one of the effective methods to prevent us from being affected by the vicious
COVID virus. Due to carelessness and unawareness of the local public towards this issue a
facemask detection system can be used by the authorities to keep a check on them and create
awareness about the same.
This project titled Face Mask Detection has been developed using a machine learning technique
known as transfer learning and by using image classification method of
MobileNetV3small .Steps followed while building the model are data collection, data
preprocessing , data splitting , model-rectification ,model testing and model implementation.
Model is capable of predicting if individuals are wearing or not wearing masks at an accuracy of
99.81 percent. We also explored using other image classification models like VGG16 with an
accuracy of 96 percent, MobileNetV2 with an accuracy of 97.21 percent and ResNet101 with an
accuracy of 97.79 percent to detect masks. We have implemented the Transfer learning method
that uses already acquired knowledge by avoiding overfitting of the models.